backend/requirements.txt
flask==2.3.1
flask-cors==3.0.10
openai==0.27.6
python-dotenv
flask-limiter
flask-caching